add const.go for all const
